+++ src/render.c
@@ -14,13 +14,13 @@ void plot_axes(window_size_t w, render_state_t* state) {
     for (int row = 0; row < w.rows; row++) draw_point(y_axis_col, row, '|');
 }
 
-void plot_function(window_size_t w, render_state_t* state, float (*f)(float),
+void plot_function(window_size_t w, render_state_t* state, double (*f)(double),
                    int delay) {
-    int left_bound = screen_to_cartesian_x(0, state->x_step, state->x_offset);
-    int right_bound =
+    double left_bound = screen_to_cartesian_x(0, state->x_step, state->x_offset);
+    double right_bound =
         screen_to_cartesian_x(w.cols, state->x_step, state->x_offset);
 
-    for (float x = left_bound; x < right_bound; x += state->x_step) {
+    for (double x = left_bound; x < right_bound; x += state->x_step) {
         draw_point(cartesian_to_screen_x(x, state->x_step, state->x_offset),
                    cartesian_to_screen_y(f(x), state->y_step, state->y_offset),
                    '*');
